ApiAnnotations : HsForAllTy discards parens
The annotations for the following code

    pTokenCostStr :: forall a .((Show a) => [a] -> Int -> String)

have detached annotations for the parens around (Show a)... String.

Trac ticket #10363

Note: This is currently a shelved attempt.

The concern is the complexity required in hs_forall_ty for the nested
case. So it should rather be dealt with when that is resolved too.